Added structure for lifter, grabber, and turntable, and fixed OI bug
The OI bug this fixed was the one where the first run of teleop after a
robot reset would cause an error because the OI would scan for joysticks
before the first driver station packet arrived. Now it repeatedly checks
for joysticks while disabled. There is now a new bug where it constantly
prints warnings, but I don't know if that can be fixed.